more gui refactoring progress
